reasoning_engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
neo4j==5.14.1
torch==2.1.0
torch-geometric==2.4.0
numpy==1.26.2
pandas==2.1.3
pydantic==2.5.2
python-dotenv==1.0.0
